Refactor vector calculations into a separate module
These are fairly generic transformations, so it makes sense to me to
keep them separate from some of our functions specific to our data
files. I'm moving them now because I may want to use those names, or
something similar, to handle transformations for wind since we need to
report (u, v) for the distribution, but want to report magnitude for the
map.
